+++ lib/cv_audio.dart
@@ -1,6 +1,6 @@
 export 'src/cv_audio_player.dart';
-export 'src/cv_audio_recorder.dart';
-export 'src/cv_audio_router.dart';
+export 'src/recorder/recorder.dart';
+export 'src/router/router.dart';
 export 'src/cv_audio_transcription.dart';
 export 'src/cv_audio_context.dart';
 export 'src/model/model.dart';