+++ lib/src/recorder/recorder.dart
@@ -1,0 +1,1 @@
+export 'src/cv_audio_recorder.dart' if (dart.library.html) 'src/cv_audio_recorder_web.dart';