+++ lib/src/utils/src/permissions/cv_audio_permissions_state.dart
@@ -16,7 +16,7 @@ class PermissionState {
   }
 
   Future<PermissionState> checkAndRequest() async {
-    if (this.isAllowed) {
+    if (!this.isAllowed) {
       permissionStatus = await permission.request();
     }
     return this;